Rudi changes: seperated out main, changed the globals to be static... so technically still globals
